+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Common
     * fix: cmd wait delay
 
     * cmd API function accepts args as a list
+    * xc/cmd moved to runtime
     * install option "--skip-venv". "--skip-check" no longer skips venv building
     * deep code audit and tons of bug fixes
     * Core Plugins API v2, new functions: